+++ tobacco-core/src/tobacco/core/components/DebuggingComponent.java
@@ -3,7 +3,7 @@ package tobacco.core.components;
 public class DebuggingComponent implements Component {
 
 	@Override
-	public int getComponentType() {
+	public short getComponentType() {
 		return DEBUGGING_C;
 	}
