+++ tobacco-core/src/tobacco/core/components/RotationComponent.java
@@ -3,7 +3,7 @@ package tobacco.core.components;
 public class RotationComponent implements Component {
 
 	@Override
-	public int getComponentType() {
+	public short getComponentType() {
 		return ROTATION_C;
 	}
