+++ tobacco-core/src/tobacco/core/components/ScaleComponent.java
@@ -3,7 +3,7 @@ package tobacco.core.components;
 public class ScaleComponent implements Component {
 
 	@Override
-	public int getComponentType() {
+	public short getComponentType() {
 		return SCALE_C;
 	}
 }
